feat: marshmallow 패키지를 이용한 스키마 추가 및 정리
